+++ v2/index_gen_test.go
@@ -23,7 +23,7 @@ func TestReadOrGenerateIndex(t *testing.T) {
 				v1, err := os.Open("testdata/sample-v1.car")
 				require.NoError(t, err)
 				defer v1.Close()
-				want, err := index.Generate(v1)
+				want, err := GenerateIndex(v1)
 				require.NoError(t, err)
 				return want
 			},
@@ -31,7 +31,7 @@ func TestReadOrGenerateIndex(t *testing.T) {
 		},
 		{
 			"CarV2WithIndexIsReturnedAsExpected",
-			"testdata/sample-v1.car",
+			"testdata/sample-wrapped-v2.car",
 			func(t *testing.T) index.Index {
 				v2, err := os.Open("testdata/sample-wrapped-v2.car")
 				require.NoError(t, err)
@@ -57,6 +57,51 @@ func TestReadOrGenerateIndex(t *testing.T) {
 			require.NoError(t, err)
 			t.Cleanup(func() { assert.NoError(t, carFile.Close()) })
 			got, err := ReadOrGenerateIndex(carFile)
+			if tt.wantErr {
+				require.Error(t, err)
+			}
+			want := tt.wantIndexer(t)
+			require.Equal(t, want, got)
+		})
+	}
+}
+
+func TestGenerateIndexFromFile(t *testing.T) {
+	tests := []struct {
+		name        string
+		carPath     string
+		wantIndexer func(t *testing.T) index.Index
+		wantErr     bool
+	}{
+		{
+			"CarV1IsIndexedAsExpected",
+			"testdata/sample-v1.car",
+			func(t *testing.T) index.Index {
+				v1, err := os.Open("testdata/sample-v1.car")
+				require.NoError(t, err)
+				defer v1.Close()
+				want, err := GenerateIndex(v1)
+				require.NoError(t, err)
+				return want
+			},
+			false,
+		},
+		{
+			"CarV2IsErrorSinceOnlyV1PayloadIsExpected",
+			"testdata/sample-wrapped-v2.car",
+			func(t *testing.T) index.Index { return nil },
+			true,
+		},
+		{
+			"CarOtherThanV1OrV2IsError",
+			"testdata/sample-rootless-v42.car",
+			func(t *testing.T) index.Index { return nil },
+			true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := GenerateIndexFromFile(tt.carPath)
 			if tt.wantErr {
 				require.Error(t, err)
 			}
